Ajustes para  template.

+++ src/optim-step3-publication.sql
@@ -60,11 +60,12 @@ FROM (
             'isGeoaddress', iif(pf.ftype_info->>'class_ftname'='geoaddress','true'::jsonb,'false'::jsonb),
         'bytes_mb', (pc.kx_profile->'publication_summary'->'bytes')::bigint / 1048576.0,
         'bytes_mb_round2', ROUND(((pc.kx_profile->'publication_summary'->'bytes')::bigint / 1048576.0),0.01),
-        'avg_density_round2', ROUND(((pc.kx_profile->'publication_summary'->'avg_density')::bigint),0.01),
-        'size_round2', ROUND(((pc.kx_profile->'publication_summary'->'size')::bigint),2),
+        'avg_density_round2', ROUND(((pc.kx_profile->'publication_summary'->'avg_density')::float),0.01),
+        'size_round2', ROUND(((pc.kx_profile->'publication_summary'->'size')::float),2),
+        
         'bytes_mb_round4', ROUND(((pc.kx_profile->'publication_summary'->'bytes')::bigint / 1048576.0),0.0001),
-        'avg_density_round4', ROUND(((pc.kx_profile->'publication_summary'->'avg_density')::bigint),0.0001),
-        'size_round4', ROUND(((pc.kx_profile->'publication_summary'->'size')::bigint),4)
+        'avg_density_round4', ROUND(((pc.kx_profile->'publication_summary'->'avg_density')::float),0.0001),
+        'size_round4', ROUND(((pc.kx_profile->'publication_summary'->'size')::float),4)
         
   ) || (pc.kx_profile->'publication_summary') AS publication_summary
